+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Change 'curl-static' version from 8.17.0 to 8.21.0.
 - AppImagePool and bauh are replaced by Shelly for a GUI package manager.
 - Change the exFAT partition to be created during the first-time setup.
+- Change the dual-boot installer to use Calamares, which also supports full disk installations.
 
 ## 4.5.0 - The Ventoy Resize Update